Force override motionPath to return motion for zero movement
When appying autolevel to gcode there is a bug where a G0 or G1 that results in no movement would be
skipped and the autolevel interpolation would not be applied. This could have terrible consequences
if the autolevel plane is far from Z=0. The fix allows for motionPath to return the movement, even
if it is zero distance on all 3 axes, so that the autoleveling can be applied.